+++ ui/src/views/index.js
@@ -1,6 +1,7 @@
 export { default as PreferencesView } from './PreferencesView';
 export { default as ClaudeSettingsView } from './ClaudeSettingsView';
 export { default as GeminiSettingsView } from './GeminiSettingsView';
+export { default as AntigravitySettingsView } from './AntigravitySettingsView';
 export { default as CreateMcpView } from './CreateMcpView';
 export { default as EnvView } from './EnvView';
 export { default as SubprojectsView } from './SubprojectsView';